% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{exposure.response.sample.binary}
\alias{exposure.response.sample.binary}
\title{Sample simulated data for exposure-response for binary endpoint}
\format{
A data frame with columns:
\describe{
\item{conc}{Simulated exposure}
\item{y_logit}{Simulated logit(p)}
\item{y_prob}{Simulated p}
\item{y}{Simulated event (1 or 0)}
\item{y_cov_logit}{Simulated logit(p) for model with covariate}
\item{y_cov_prob}{Simulated p for model with covariate}
\item{y_cov}{Simulated event (1 or 0) for model with covariate}
\item{sex_num}{1 or 0}
\item{sex}{female or male}
}
}
\usage{
exposure.response.sample.binary
}
\description{
Sample simulated data for exposure-response for binary endpoint
}
\examples{
exposure.response.sample
}
\keyword{datasets}
