% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{exposure.response.sample}
\alias{exposure.response.sample}
\title{Sample simulated data for exposure-response.}
\format{
A data frame with columns:
\describe{
\item{dose}{Dose levels used for simulation of pharmacokinetics}
\item{exposure}{Simulated exposure}
\item{response}{Simulated pharmacodynamic response}
}
}
\usage{
exposure.response.sample
}
\description{
Sample simulated data for exposure-response.
}
\examples{
exposure.response.sample
}
\keyword{datasets}
