% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{exposure.response.sample.test}
\alias{exposure.response.sample.test}
\title{Sample simulated data for exposure-response with covariates for package testing}
\format{
A data frame with columns:
\describe{
\item{dose}{Dose levels used for simulation of pharmacokinetics}
\item{conc}{Simulated exposure}
\item{resp}{Simulated pharmacodynamic response}
\item{cov1}{Covariate 1 for e0}
\item{cov2}{Covariate 2 for emax}
\item{cov3}{Covariate 3 for ec50 (data type factor)}
\item{cov3num}{Covariate 3 for ec50 (data type numeric)}
}
}
\usage{
exposure.response.sample.test
}
\description{
Sample simulated data for exposure-response with covariates for package testing
}
\examples{
exposure.response.sample
}
\keyword{datasets}
