% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param}
\alias{extract_param}
\title{Extract posterior draws of key parameters}
\usage{
extract_param(object)
}
\arguments{
\item{object}{A \code{stanemax} class object}
}
\value{
A tibble containing posterior draws of key parameters.
If covariate(s) are included in the model, posterior draws for different combinations
of covariates are supplied in a long format - e.g. if there are posterior draws of 100 samples
and 4 levels of the covariates, the returned tibble will have the length of 400
}
\description{
Extract posterior draws of key parameters
}
