% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_stan.R
\name{lm_stan_formula}
\alias{lm_stan_formula}
\title{Bayesian linear regression with Stan using formula notation}
\usage{
lm_stan_formula(formula, data, ...)
}
\arguments{
\item{formula}{a symbolic description of variables for Emax model fit.}

\item{data}{an optional data frame containing the variables in the model.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian linear regression with Stan using formula notation
}
\examples{
df <- data.frame(a = rnorm(10), b = rnorm(10))
fit <- lm_stan_formula(b ~ a, data = df)
print(fit)

\dontrun{

}

}
