% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\alias{log_lik.stanemax}
\alias{log_lik.stanemaxbin}
\title{Calculate log-likelihoods from posterior samples. Data can be either
original data used for model fit or new data.}
\usage{
log_lik(object, ...)

\method{log_lik}{stanemax}(object, newdata = NULL, ...)

\method{log_lik}{stanemaxbin}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{stanemax} or \code{stanemaxbin} object}

\item{...}{Currently unused arguments}

\item{newdata}{New data used for prediction. If NULL, original data is used.}
}
\value{
\eqn{S} by \eqn{N} matrix of log-likelihoods, where each row
corresponds to a draw from the posterior distribution and each column corresponds to a data point.
}
\description{
See \link[rstantools:log_lik]{rstantools::log_lik} for more details.
}
