% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanemax-package.R
\docType{package}
\name{rstanemax-package}
\alias{rstanemax-package}
\alias{rstanemax}
\title{The 'rstanemax' package.}
\description{
Perform sigmoidal Emax model fit using Stan without writing Stan model code.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
