% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanemax-methods.R
\name{stanemax-methods}
\alias{stanemax-methods}
\alias{print.stanemax}
\alias{plot.stanemax}
\title{Methods for stanemax objects}
\usage{
\method{print}{stanemax}(object, ...)

\method{plot}{stanemax}(object, show.ci = TRUE, show.pi = FALSE, ...)
}
\arguments{
\item{object}{An object of class `stanemax`}

\item{show.ci}{An logical specifying if the output figure include
credible interval of posterior prediction. Default TRUE.}

\item{show.pi}{An logical specifying if the output figure include
prediction interval. Default FALSE.}
}
\description{
Methods for stanemax objects
}
