% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanemax-methods.R
\name{as_draws}
\alias{as_draws}
\alias{as_draws_list.stanemax}
\alias{as_draws_array.stanemax}
\alias{as_draws_df.stanemax}
\alias{as_draws_matrix.stanemax}
\alias{as_draws_rvars.stanemax}
\title{Convert stanemax object to a posterior draws object}
\usage{
\method{as_draws_list}{stanemax}(x, ...)

\method{as_draws_array}{stanemax}(x, ...)

\method{as_draws_df}{stanemax}(x, ...)

\method{as_draws_matrix}{stanemax}(x, ...)

\method{as_draws_rvars}{stanemax}(x, ...)
}
\arguments{
\item{x}{An object of class stanemax}

\item{...}{Arguments passed to next method}
}
\value{
A draws object of the appropriate subclass
}
\description{
Convert stanemax object to a posterior draws object
}
\examples{
data(exposure.response.sample)
fit <- stan_emax(response ~ exposure, exposure.response.sample)
if(require(posterior, quietly = TRUE)) {
  posterior::as_draws_list(fit)
  posterior::as_draws_array(fit)
  posterior::as_draws_df(fit)
  posterior::as_draws_matrix(fit)
  posterior::as_draws_rvars(fit)
}
}
