% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanemax-package.R
\docType{package}
\name{rstanemax-package}
\alias{rstanemax-package}
\alias{rstanemax}
\title{The 'rstanemax' package.}
\description{
Perform sigmoidal Emax model fit using Stan without writing Stan model code.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yoshidk6/rstanemax}
  \item \url{https://yoshidk6.github.io/rstanemax/}
  \item Report bugs at \url{https://github.com/yoshidk6/rstanemax/issues}
}

}
\author{
\strong{Maintainer}: Kenta Yoshida \email{yoshida.kenta.6@gmail.com} (\href{https://orcid.org/0000-0003-4967-3831}{ORCID})

Authors:
\itemize{
  \item Danielle Navarro \email{djnavarro@protonmail.com} (\href{https://orcid.org/0000-0001-7648-6578}{ORCID})
}

Other contributors:
\itemize{
  \item Trustees of Columbia University (src/init.cpp, tools/make_cc.R, R/stanmodels.R) [copyright holder]
}

}
