% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_emax.R
\name{stan_emax}
\alias{stan_emax}
\alias{stan_emax_run}
\title{Bayesian Emax model fit with Stan}
\usage{
stan_emax(formula, data, gamma.fix = 1, e0.fix = NULL, priors = NULL,
  ...)

stan_emax_run(stanmodel, standata, ...)
}
\arguments{
\item{formula}{a symbolic description of variables for Emax model fit.}

\item{data}{an optional data frame containing the variables in the model.}

\item{gamma.fix}{a numeric or NULL to specify gamma (Hill coefficient) in the sigmoidal Emax model.
If NULL, gamma will be estimated from the data.
If numeric, gamma is fixed at the number provided.
Default = 1 (normal Emax model).}

\item{e0.fix}{a numeric or NULL to specify E0 in the Emax model.
If NULL, E0 will be estimated from the data.
If numeric, E0 is fixed at the number provided.
Default = NULL (estimate from the data).}

\item{priors}{a list specifying priors of parameters.
Not implemented yet.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{stanmodel}{a Stan model object.}

\item{standata}{a data file for model fit}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanemax`

An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Add explanations

Add explanations
}
\examples{
data(exposure.response.sample)
ggplot2::qplot(exposure, response, data = exposure.response.sample)

df <- data.frame(a = rnorm(10), b = rnorm(10))
fit <- lm_stan_formula(b ~ a, data = df)
print(fit)

\dontrun{
data(exposure.response.sample)
fit <- stan_emax(response ~ exposure, data = exposure.response.sample)

}

}
