% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanemax-methods.R
\name{stanemax-methods}
\alias{stanemax-methods}
\alias{print.stanemax}
\alias{plot.stanemax}
\title{Methods for stanemax objects}
\usage{
\method{print}{stanemax}(object)

\method{plot}{stanemax}(object, show.ci = TRUE, show.pi = FALSE)
}
\description{
Methods for stanemax objects
}
