% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanemax-methods.R
\name{stanemax-methods}
\alias{stanemax-methods}
\alias{print.stanemax}
\alias{print.stanemaxbin}
\alias{extract_stanfit}
\alias{extract_obs_mod_frame}
\alias{plot.stanemax}
\alias{plot.stanemaxbin}
\title{Methods for stanemax objects}
\usage{
\method{print}{stanemax}(x, digits_summary = 2, ...)

\method{print}{stanemaxbin}(x, digits_summary = 2, ...)

extract_stanfit(x)

extract_obs_mod_frame(x)

\method{plot}{stanemax}(x, show.ci = TRUE, show.pi = FALSE, ci = 0.9, pi = 0.9, ...)

\method{plot}{stanemaxbin}(x, show.ci = TRUE, show.pi = FALSE, ci = 0.9, pi = 0.9, ...)
}
\arguments{
\item{x}{An object of class \code{stanemax} or \code{stanemaxbin}}

\item{digits_summary}{The number of significant digits to use when printing
the summary, defaulting to 2. Applies to the quantities other than the
effective sample size, which is always rounded to the nearest integer.}

\item{...}{Additional arguments passed to methods.}

\item{show.ci}{An logical specifying if the output figure include
credible interval of posterior prediction. Default TRUE.}

\item{show.pi}{An logical specifying if the output figure include
prediction interval. Default FALSE.}

\item{ci}{Credible interval range.}

\item{pi}{Prediction interval range.}
}
\description{
Methods for stanemax objects
}
